+++ deps/zones.ts
@@ -1,2 +1,1 @@
-// export * from "https://deno.land/x/zones@v0.1.0-beta.12/mod.ts";
-export * from "../../zones/mod.ts";
+export * from "https://deno.land/x/zones@v0.1.0-beta.12/mod.ts";
