+++ examples/transfer.ts
@@ -4,15 +4,15 @@ import * as U from "../util/mod.ts";
 
 const root = C.extrinsic({
   client: T.westend,
-  sender: C.MultiAddress.fromKeypair(T.alice),
+  sender: C.compat.multiAddressFromKeypair(T.alice),
   palletName: "Balances",
   methodName: "transfer",
   args: {
     value: 12345n,
-    dest: C.MultiAddress.fromKeypair(T.bob),
+    dest: C.compat.multiAddressFromKeypair(T.bob),
   },
 })
-  .signed(C.Signer.fromKeypair(T.alice))
+  .signed(C.compat.signerFromKeypair(T.alice))
   .watch(function(status) {
     console.log(status);
     if (C.TransactionStatus.isTerminal(status)) {
