+++ mod.ts
@@ -1,9 +1,17 @@
+export * as compat from "./compat/mod.ts";
 export * as $ from "./deps/scale.ts";
 export { BitSequence } from "./deps/scale.ts";
 export * as Z from "./deps/zones.ts";
 export * from "./effects/mod.ts";
 export * as M from "./frame_metadata/mod.ts";
-export { $era, $null, ChainError, type Era, MultiAddress, Signer } from "./frame_metadata/mod.ts";
+export {
+  $era,
+  $null,
+  ChainError,
+  type Era,
+  MultiAddress,
+  type Signer,
+} from "./frame_metadata/mod.ts";
 export { kusama, moonbeam, polkadot, rococo, TransactionStatus, westend } from "./known/mod.ts"; // TODO: get rid of this!
 export * as rpc from "./rpc/mod.ts";
 export { contramapListener, hex, type Listener } from "./util/mod.ts";
